requirements.txt
farm-haystack
fastapi
qdrant-haystack==1.0.10
farm-haystack[inference]
uvicorn
